requirements.txt
openwakeword
sounddevice
numpy
whispercpp
picamera2
httpx
python-dotenv
piper-tts
